unify `bin_helpers` and `dylib` utility modules
